# 🚀 Fix: Achieve 90+ Lighthouse Score - Resolve Font 404 & CSP Violations