--- a/DEPLOYMENT_CHECKLIST.md
+++ b/DEPLOYMENT_CHECKLIST.md
@@ -1,0 +1,216 @@
+# 🚀 Deployment Checklist - Performance Fixes
+
+## Pre-Deployment Validation ✅
+
+- [x] Run validation script: `pnpm perf:validate`
+- [x] All validation checks passed
+- [x] No local font files present
+- [x] CSP headers configured correctly
+- [x] Font optimization in place
+- [x] Image caching headers added
+- [x] Hero image (LCP) exists
+
+## Deployment Steps
+
+### 1. Commit Changes
+```bash
+git add .
+git commit -m "fix: resolve font 404 and CSP violations for 90+ Lighthouse score"
+git push origin main
+```
+
+### 2. Monitor Deployment
+- [ ] Go to Netlify dashboard
+- [ ] Watch build logs for errors
+- [ ] Wait for deployment to complete
+- [ ] Note the deployment URL
+
+### 3. Post-Deployment Verification
+
+#### A. Console Errors Check
+- [ ] Open deployed site in Chrome
+- [ ] Open DevTools (F12)
+- [ ] Go to Console tab
+- [ ] Verify: **Zero errors** ✅
+- [ ] Verify: No 404 for fonts ✅
+- [ ] Verify: No CSP violations ✅
+
+#### B. Issues Panel Check
+- [ ] Open DevTools
+- [ ] Go to Issues panel
+- [ ] Verify: **Zero issues** ✅
+- [ ] Verify: No CSP warnings ✅
+- [ ] Verify: No security warnings ✅
+
+#### C. Network Tab Check
+- [ ] Open DevTools Network tab
+- [ ] Reload page
+- [ ] Check font files load successfully
+- [ ] Check images use WebP/AVIF
+- [ ] Verify cache headers are set
+
+### 4. Lighthouse Audit
+
+#### Desktop Audit
+- [ ] Open DevTools
+- [ ] Go to Lighthouse tab
+- [ ] Select "Desktop"
+- [ ] Select all categories
+- [ ] Click "Generate report"
+- [ ] **Performance**: Target 90+ ⭐
+- [ ] **Best Practices**: Target 95+ ⭐
+- [ ] **SEO**: Target 95+ ⭐
+- [ ] **Accessibility**: Target 95+ ⭐
+
+#### Mobile Audit
+- [ ] Select "Mobile"
+- [ ] Generate report
+- [ ] **Performance**: Target 90+ ⭐
+- [ ] **Best Practices**: Target 95+ ⭐
+- [ ] **SEO**: Target 95+ ⭐
+- [ ] **Accessibility**: Target 95+ ⭐
+
+### 5. Core Web Vitals Check
+
+- [ ] **LCP (Largest Contentful Paint)**: < 2.5s
+- [ ] **FID/INP (Interaction)**: < 200ms
+- [ ] **CLS (Cumulative Layout Shift)**: < 0.1
+
+### 6. Functional Testing
+
+- [ ] Homepage loads correctly
+- [ ] Navigation works
+- [ ] Images display properly
+- [ ] Fonts render correctly
+- [ ] Stripe checkout works
+- [ ] Forms submit successfully
+- [ ] Mobile responsive
+
+## Expected Results
+
+### Before Fixes:
+- ❌ Font 404 errors in console
+- ❌ CSP violations for Netlify
+- ⚠️ Performance score: ~70-80
+- ⚠️ Best Practices: ~80-85
+
+### After Fixes:
+- ✅ Zero console errors
+- ✅ Zero CSP violations
+- ✅ Performance score: 90+
+- ✅ Best Practices: 95+
+
+## Troubleshooting
+
+### If scores are still below 90:
+
+1. **Clear Cache**
+   ```bash
+   # In Chrome DevTools
+   Right-click Reload → Empty Cache and Hard Reload
+   ```
+
+2. **Check Specific Issues**
+   - Look at Lighthouse report details
+   - Check "Opportunities" section
+   - Review "Diagnostics" section
+
+3. **Verify Deployment**
+   ```bash
+   # Check if latest changes are deployed
+   curl -I https://your-site.netlify.app
+   ```
+
+4. **Run Local Audit**
+   ```bash
+   pnpm perf:audit
+   ```
+
+### If console errors persist:
+
+1. **Check Network Tab**
+   - Look for failed requests
+   - Check request URLs
+   - Verify response codes
+
+2. **Check CSP Headers**
+   ```bash
+   curl -I https://your-site.netlify.app | grep -i "content-security"
+   ```
+
+3. **Verify Middleware**
+   - Check middleware.ts is deployed
+   - Verify CSP directives are correct
+
+## Success Metrics
+
+### Immediate (Post-Deployment):
+- [ ] Zero console errors
+- [ ] Zero CSP violations
+- [ ] Zero issues in DevTools
+- [ ] All resources load successfully
+
+### Performance (Lighthouse):
+- [ ] Performance: 90+
+- [ ] Best Practices: 95+
+- [ ] SEO: 95+
+- [ ] Accessibility: 95+
+
+### Core Web Vitals:
+- [ ] LCP: Green (< 2.5s)
+- [ ] FID/INP: Green (< 200ms)
+- [ ] CLS: Green (< 0.1)
+
+## Documentation
+
+- **Summary**: `PERFORMANCE_FIXES_SUMMARY.md`
+- **Detailed Fixes**: `docs/PERFORMANCE_FIXES_90_PLUS.md`
+- **Performance Plan**: `PERFORMANCE_PLAN.md`
+
+## Next Actions After Success
+
+1. **Monitor Production**
+   - Set up Google Analytics
+   - Track Core Web Vitals
+   - Monitor real user metrics
+
+2. **Set Up Alerts**
+   - Performance degradation alerts
+   - Error monitoring
+   - Uptime monitoring
+
+3. **Continuous Optimization**
+   - Regular Lighthouse audits
+   - Performance budgets
+   - A/B testing
+
+## Quick Commands
+
+```bash
+# Validate configuration
+pnpm perf:validate
+
+# Run performance audit
+pnpm perf:audit
+
+# Build and audit
+pnpm perf:build
+
+# Deploy
+git push origin main
+```
+
+## Contact & Support
+
+If you encounter issues:
+1. Check troubleshooting section above
+2. Review Lighthouse report details
+3. Check browser console for specific errors
+4. Verify all files were deployed correctly
+
+---
+
+**Ready to Deploy**: ✅ Yes
+**Validation Status**: ✅ Passed
+**Expected Outcome**: 90+ Lighthouse scores across all metrics
+