--- a/netlify.toml
+++ b/netlify.toml
@@ -43,6 +43,30 @@ package = "@netlify/plugin-nextjs"
   for = "/images/*"
   [headers.values]
     Cache-Control = "public, max-age=31536000, immutable"
+
+# Font files caching
+[[headers]]
+  for = "*.woff2"
+  [headers.values]
+    Cache-Control = "public, max-age=31536000, immutable"
+    Access-Control-Allow-Origin = "*"
+
+[[headers]]
+  for = "*.woff"
+  [headers.values]
+    Cache-Control = "public, max-age=31536000, immutable"
+    Access-Control-Allow-Origin = "*"
+
+# WebP and AVIF images
+[[headers]]
+  for = "*.webp"
+  [headers.values]
+    Cache-Control = "public, max-age=31536000, immutable"
+
+[[headers]]
+  for = "*.avif"
+  [headers.values]
+    Cache-Control = "public, max-age=31536000, immutable"
     
 # API-specific headers
 [[headers]]
@@ -65,7 +89,10 @@ package = "@netlify/plugin-nextjs"
 [[headers]]
   for = "/*"
   [headers.values]
-    X-Frame-Options = "DENY"
+    X-Frame-Options = "SAMEORIGIN"
     X-Content-Type-Options = "nosniff"
+    X-XSS-Protection = "1; mode=block"
     Referrer-Policy = "strict-origin-when-cross-origin"
-    Permissions-Policy = "payment=*, geolocation=(), microphone=(), camera=()"
+    Permissions-Policy = "payment=(self \"https://*.stripe.com\"), geolocation=(), microphone=(), camera=()"
+    Content-Security-Policy = "default-src 'self'; script-src 'self' 'unsafe-inline' 'unsafe-eval' *.stripe.com *.supabase.co js.stripe.com https://www.googletagmanager.com https://www.google-analytics.com; style-src 'self' 'unsafe-inline' fonts.googleapis.com; img-src 'self' data: blob: *.stripe.com *.supabase.co https://www.google-analytics.com; font-src 'self' fonts.gstatic.com data:; connect-src 'self' *.stripe.com *.supabase.co api.stripe.com https://www.google-analytics.com https://analytics.google.com; frame-src 'self' *.stripe.com *.netlify.com *.netlify.app https://app.netlify.com; object-src 'none'; base-uri 'self'; form-action 'self'; upgrade-insecure-requests"
+    Strict-Transport-Security = "max-age=31536000; includeSubDomains; preload"