--- a/next.config.js
+++ b/next.config.js
@@ -11,12 +11,15 @@ const nextConfig = {
   // - optimizePackageImports (can cause dependency resolution issues)
   // experimental: {},
 
-  // Essential image configuration  
+  // Essential image configuration
   images: {
     domains: ['venomappdevelopment.com'],
     formats: ['image/webp', 'image/avif'],
     dangerouslyAllowSVG: true,
     contentSecurityPolicy: "default-src 'self'; script-src 'none'; sandbox;",
+    deviceSizes: [640, 750, 828, 1080, 1200, 1920, 2048, 3840],
+    imageSizes: [16, 32, 48, 64, 96, 128, 256, 384],
+    minimumCacheTTL: 31536000,
   },
 
   // Core compiler settings (stable features only)