@gpb360 2025-10-07
@claude this actually made the performance go down can we inbvestigate why